+++ CHANGELOG.md
@@ -62,6 +62,9 @@
     the domain was found, or `$false` if it was not.
 - Changes to ADUser
   - Remove unused non-mandatory parameters from the Get-TargetResource ([issue #293](https://github.com/PowerShell/ActiveDirectoryDsc/issues/293)).
+- Changes to ADDomainController
+  - Add InstallDns parameter to enable promotion without installing local
+    DNS Server Service ([issue #87](https://github.com/PowerShell/xActiveDirectory/issues/87)).
 
 ## 4.0.0.0